Move replacement strategy closer to usage
