+++ src/refactorings/extract/extract-generic-type/extract-generic-type.test.ts
@@ -4,7 +4,7 @@ import { InMemoryEditor } from "../../../editor/adapters/in-memory-editor";
 import { testEach } from "../../../tests-helpers";
 
 import { extractGenericType } from "./extract-generic-type";
-import { ReplacementStrategy } from "../../../replacement-strategy";
+import { ReplacementStrategy } from "../replacement-strategy";
 import { Position } from "../../../editor/position";
 
 describe("Extract Generic Type", () => {
