+++ src/refactorings/extract/extract-variable/extract-variable.multiple-occurrences.test.ts
@@ -4,7 +4,7 @@ import { Position } from "../../../editor/position";
 import { InMemoryEditor } from "../../../editor/adapters/in-memory-editor";
 
 import { extractVariable } from "./extract-variable";
-import { ReplacementStrategy } from "../../../replacement-strategy";
+import { ReplacementStrategy } from "../replacement-strategy";
 import { testEach } from "../../../tests-helpers";
 
 describe("Extract Variable - Multiple occurrences", () => {
